migrate with-mongodb-mongoose to app dir
