+++ examples/with-mongodb-mongoose/package.json
@@ -9,8 +9,9 @@
     "@types/node": "^20.4.7",
     "@types/react": "^18.2.18",
     "@types/react-dom": "^18.2.7",
+    "eslint-config-next": "^14.0.3",
     "mongoose": "^7.4.2",
-    "next": "latest",
+    "next": "^14.0.3",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "swr": "^2.2.0",
