+++ examples/with-mongodb-mongoose/src/app/new/page.tsx
@@ -1,6 +1,6 @@
-import Form from '../components/Form'
+import { Form } from '@/components/Form'
 
-const NewPet = () => {
+export default function NewPet() {
   const petForm = {
     name: '',
     owner_name: '',
@@ -15,5 +15,3 @@ const NewPet = () => {
 
   return <Form formId="add-pet-form" petForm={petForm} />
 }
-
-export default NewPet
