+++ examples/with-mongodb-mongoose/src/models/Pet.ts
