+++ examples/with-mongodb-mongoose/tsconfig.json
@@ -16,8 +16,13 @@
     "incremental": true,
     "paths": {
       "@/*": ["./src/*", "./lib/*"]
-    }
+    },
+    "plugins": [
+      {
+        "name": "next"
+      }
+    ]
   },
-  "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx"],
+  "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx", ".next/types/**/*.ts"],
   "exclude": ["node_modules"]
 }
